pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jcapy"
version = "4.1.8"
description = "JCapy CLI - One-Army Orchestrator"
authors = [{name = "Irfan Ali", email = "nazrijz336@gmail.com"}]
requires-python = ">=3.11"
dependencies = [
    "rich>=13.0.0",
    "mcp>=1.2.0",
    "chromadb>=0.4.0",
    "posthog>=3.0.0",
    "textual>=0.50.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "grpcio>=1.78.1",
    "grpcio-tools>=1.78.1",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
]
readme = "README.md"
keywords = ["cli", "developer-tools", "automation", "productivity", "terminal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

[project.urls]
Homepage = "https://github.com/ponli550/JCapy"
Repository = "https://github.com/ponli550/JCapy.git"
Issues = "https://github.com/ponli550/JCapy/issues"

[project.optional-dependencies]
pro = ["pinecone>=3.0.0"]
web = [
    "pyzmq>=25.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]
all = [
    "pinecone>=3.0.0",
    "pyzmq>=25.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
jcapy = "jcapy.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true


[project.entry-points."jcapy.plugins"]
# External plugins register here
